Ted fix 376-high-vulnerability

+++ includes/gdpr-integration/gdpr-user-profile.php
@@ -59,23 +59,28 @@ class WPAS_GDPR_User_Profile {
 
 		if ( isset( $_GET['file'] ) && isset( $_GET['check'] ) ) {
 			$nonce = ( ! empty( $_GET['check'] ) ) ? sanitize_text_field( $_GET['check'] ) : '';
-			if ( wp_verify_nonce( $nonce, 'as-validate-download-url' ) ) {
+			if ( wp_verify_nonce( $nonce, 'as-validate-download-url' )  && ( intval( $_GET['file'] ) == get_current_user_id() ) ) {
 				$user = intval( $_GET['file'] );
 				if ( ! $this->user_export_dir ) {
 					$this->user_export_dir = $this->set_log_dir( $user );
 				}
-				header( 'Content-Description: File Transfer' );
-				header( 'Content-type: application/zip' );
-				header( 'Content-Disposition: attachment; filename=exported-data.zip' );
-				header( 'Content-length: ' . filesize( $this->user_export_dir . '/exported-data.zip' ) );
-				header( 'Pragma: no-cache' );
-				header( 'Expires: 0' );
-				readfile( $this->user_export_dir . '/exported-data.zip' );
-				if (!unlink($this->user_export_dir . '/exported-data.zip') ){
-					return new WP_Error( 'file_deleting_error', __( 'Error deleting ' . $this->user_export_dir . '/exported-data.zip', 'awesome-support' ) );
+				if ( file_exists( $this->user_export_dir . '/exported-data.zip' ) ) {
+					header( 'Content-Description: File Transfer' );
+					header( 'Content-type: application/zip' );
+					header( 'Content-Disposition: attachment; filename=exported-data.zip' );
+					header( 'Content-length: ' . filesize( $this->user_export_dir . '/exported-data.zip' ) );
+					header( 'Pragma: no-cache' );
+					header( 'Expires: 0' );
+					readfile( $this->user_export_dir . '/exported-data.zip' );
+					if (!unlink($this->user_export_dir . '/exported-data.zip') ){
+						return new WP_Error( 'file_deleting_error', __( 'Error deleting ' . $this->user_export_dir . '/exported-data.zip', 'awesome-support' ) );
+					}
+					if (!unlink($this->user_export_dir . '/export-data.xml') ){
+						return new WP_Error( 'file_deleting_error', __( 'Error deleting ' . $this->user_export_dir . '/export-data.xml', 'awesome-support' ) );
+					}
 				}
-				if (!unlink($this->user_export_dir . '/export-data.xml') ){
-					return new WP_Error( 'file_deleting_error', __( 'Error deleting ' . $this->user_export_dir . '/export-data.xml', 'awesome-support' ) );
+				else {
+					return new WP_Error( 'security_error', __( 'Request not identified, Invalid request', 'awesome-support' ) );
 				}
 			} else {
 				return new WP_Error( 'security_error', __( 'Request not identified, Invalid request', 'awesome-support' ) );
